+++ data/__init__.py
@@ -1,3 +1,6 @@
 #!/usr/bin/env python
 
 from .db import *
+from .build_repository import *
+from .professions_repository import *
+from .specializations_repository import *
