+++ model/dao/__init__.py
@@ -4,3 +4,7 @@ from .build import *
 from .weapon import *
 from .profession import *
 from .specialization import *
+from .skill import *
+from .item import *
+from .fact import *
+from .item_stats import *
